:building_construction: Modify program structure
